added exception parts of application

+++ src/main/java/com/example/ca_cert/businesslogic/DataBaseUtility.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface DataBaseUtility extends JpaRepository<UserData,Integer> {
+public interface DataBaseUtility extends JpaRepository<UserData,Long> {
 
 }
