+++ src/main/java/com/example/ca_cert/data/UserData.java
@@ -21,5 +21,5 @@ public class UserData {
 
     @Id
     @Column(name = "aadharcard")
-    private Integer aadharcard;
+    private Long aadharcard;
 }
